Revert "feat: refactor search parameter validation into dedicated functions"
This reverts commit f648725a3da383603f71d5885983567bc137bdc2.
